pyproject.toml
[project]
name = "booker"
version = "0.1.0"
description = "Telegram bot for automation"
authors = [
    {name = "Artem Muradyan", email = "muradyanartem2@gmail.com"},
]
dependencies = [
    "google-api-python-client>=2.47.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.5.1",
    "telethon>=1.24.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
[project.optional-dependencies]

[tool.pdm]

[tool.pdm.scripts]
    start = {cmd = ["python", "-m", "booker"]}

[tool.pdm.dev-dependencies]
dev = [
    "isort>=5.10.1",
    "black>=22.3.0",
    "flake8>=4.0.1",
    "mypy>=0.950",
]
[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
